requirements-test.txt
# ──────────────────────────────────────────────────────────────────
# WordWeaveWeb — dépendances pour les tests unitaires
# Installation : pip install -r requirements-test.txt
# ──────────────────────────────────────────────────────────────────

# Framework de tests
pytest>=7.4
pytest-asyncio>=0.23
pytest-mock>=3.12

# Auth (service_c)
bcrypt>=4.0
PyJWT>=2.8
pydantic>=2.0

# ML / Embeddings (service_b)
numpy>=1.24
scikit-learn>=1.3

# Base de données (shared — mocké, mais nécessaire à l'import)
sqlalchemy>=2.0
asyncpg>=0.29

# LLM client (service_b — mocké, mais nécessaire à l'import)
openai>=1.0
